# contracts/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "ErrorResponse" : {
        "description" : "Error response",
        "type" : "object",
        "required" : [ "code" ],
        "properties" : {
          "code" : {
            "description" : "Error code",
            "type" : "string",
            "$ref" : "#/components/schemas/ErrorCode"
          },
          "errorDetails" : {
            "description" : "Additional details",
            "type" : "object",
            "$ref" : "#/components/schemas/ErrorDetails"
          }
        }
      },
...
      "MinRole" : {
        "type" : "string",
        "enum" : [ "NOT_MEMBER", "MEMBER", "ORGANIZER", "ADMIN" ]
      },
...
      "ReorderPagesRequest" : {
        "description" : "Request to reorder team pages",
        "type" : "object",
        "required" : [ "pageIds" ],
        "properties" : {
          "pageIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "description" : "Ordered list of page IDs"
          }
        }
      },
...
      "TeamListResponse" : {
        "description" : "Paginated team list response",
        "type" : "object",
        "required" : [ "teams", "total", "page", "size" ],
        "properties" : {
          "teams" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TeamDetailDto"
            },
            "description" : "List of teams"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Total number of teams"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Current page number"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Page size"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "openIdConnectUrl" : "http://localhost:8180/realms/quarkus/.well-known/openid-configuration",
        "description" : "Authentication"
      }
    }
  },
...
  "paths" : {
...
    "/api/teams" : {
      "get" : {
        "summary" : "List public teams",
        "description" : "Get a paginated list of public teams with optional search",
        "operationId" : "listTeams",
        "tags" : [ "Teams" ],
        "parameters" : [ {
          "description" : "Minimum role in team",
          "name" : "minRole",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/MinRole"
          }
        }, {
          "description" : "Page number (0-indexed)",
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "description" : "Search query to filter teams by name",
          "name" : "search",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Page size",
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Teams retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamListResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create team",
        "description" : "Create a new team. The current user will be set as the team owner.",
        "operationId" : "createTeam",
        "tags" : [ "Teams" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TeamRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Team created successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamDetailDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
    "/api/teams/{slug}" : {
      "put" : {
        "summary" : "Update team",
        "description" : "Update team information. Requires ADMIN role.",
        "operationId" : "updateTeam",
        "tags" : [ "Teams" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TeamRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Team updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamDetailDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not a team admin",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      },
      "get" : {
        "summary" : "Get team by slug",
        "description" : "Get detailed team information by URL slug",
        "operationId" : "getTeam",
        "tags" : [ "Teams" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Team retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamDetailDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Team is private and user is not a member",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Delete team",
        "description" : "Soft delete a team. Requires OWNER role.",
        "operationId" : "deleteTeam",
        "tags" : [ "Teams" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Team deleted successfully"
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not the team owner",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
...
    "/api/teams/{slug}/pages" : {
      "get" : {
        "summary" : "List team pages",
        "description" : "Get list of additional pages for a team",
        "operationId" : "listPages",
        "tags" : [ "Team Pages" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Pages retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TeamPageSummaryDto"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create page",
        "description" : "Create a new team page. Requires admin permissions. Maximum 3 additional pages per team.",
        "operationId" : "createPage",
        "tags" : [ "Team Pages" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TeamPageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Page created successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamPageDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request or maximum pages reached",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not a team admin",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
    "/api/teams/{slug}/pages/reorder" : {
      "put" : {
        "summary" : "Reorder pages",
        "description" : "Reorder team pages. Requires admin permissions.",
        "operationId" : "reorderPages",
        "tags" : [ "Team Pages" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReorderPagesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Pages reordered successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TeamPageSummaryDto"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid page IDs",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not a team admin",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
    "/api/teams/{slug}/pages/{pageSlug}" : {
      "put" : {
        "summary" : "Update page",
        "description" : "Update page information. Requires admin permissions.",
        "operationId" : "updatePage",
        "tags" : [ "Team Pages" ],
        "parameters" : [ {
          "description" : "Page URL slug",
          "name" : "pageSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TeamPageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Page updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamPageDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not a team admin",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team or page not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      },
      "get" : {
        "summary" : "Get page details",
        "description" : "Get detailed page information",
        "operationId" : "getPage",
        "tags" : [ "Team Pages" ],
        "parameters" : [ {
          "description" : "Page URL slug",
          "name" : "pageSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Page retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamPageDto"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not authorized to view this page",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team or page not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Delete page",
        "description" : "Delete a team page. Requires admin permissions.",
        "operationId" : "deletePage",
        "tags" : [ "Team Pages" ],
        "parameters" : [ {
          "description" : "Page URL slug",
          "name" : "pageSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Page deleted successfully"
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not a team admin",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team or page not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
...
    "/api/teams/{slug}/slug" : {
      "patch" : {
        "summary" : "Change team slug",
        "description" : "Change team URL slug. Requires ADMIN role.",
        "operationId" : "changeTeamSlug",
        "tags" : [ "Teams" ],
        "parameters" : [ {
          "description" : "Current team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SlugChangeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Slug changed successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamDetailDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid slug format",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not a team admin",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Slug already in use",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
...
  },
  "info" : {
    "title" : "Tribly API",
    "version" : "1.0.0",
    "description" : "API for Tribly Cycling Team Management Platform"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}